actions: setup a basic ci workflow
for now, all this does is run RSpec. we use the docker-compose environment we're
also using in test.

we run the database migrations manually as a step, but that feels like a bit of
a hack, since it means we're running them there and in the entrypoint
script. but at least this works.